+++ packages/sdk-metrics/src/aggregator/LastValue.ts
@@ -106,14 +106,17 @@ export class LastValueAggregator implements Aggregator<LastValueAccumulation> {
       descriptor,
       aggregationTemporality,
       dataPointType: DataPointType.GAUGE,
-      dataPoints: accumulationByAttributes.map(([attributes, accumulation]) => {
-        return {
-          attributes,
-          startTime: accumulation.startTime,
-          endTime,
-          value: accumulation.toPointValue(),
-        };
-      }),
+      dataPoints: accumulationByAttributes.map(
+        ([attributes, accumulation, exemplars]) => {
+          const dp = {
+            attributes,
+            startTime: accumulation.startTime,
+            endTime,
+            value: accumulation.toPointValue(),
+          };
+          return exemplars?.length ? { ...dp, exemplars } : dp;
+        }
+      ),
     };
   }
 }