+++ CHANGELOG.md
@@ -14,7 +14,11 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 
 ### :rocket: Features
 
+* feat(sdk-metrics): wire exemplar support into metrics pipeline [#6483](https://github.com/open-telemetry/opentelemetry-js/pull/6483) @CharlieTLe
+
 ### :bug: Bug Fixes
+
+* fix(sdk-metrics): fix mutation bug in ExemplarBucket.collect() that corrupted shared attributes [#6483](https://github.com/open-telemetry/opentelemetry-js/pull/6483) @CharlieTLe
 
 * fix(opentelemetry-instrumentation): improve `_warnOnPreloadedModules` function not to show warning logs when the module is not marked as loaded [#6095](https://github.com/open-telemetry/opentelemetry-js/pull/6095) @rlj1202
 * fix(sdk-trace-base): derive internal `SpanOptions` from API type to prevent drift [#6478](https://github.com/open-telemetry/opentelemetry-js/pull/6478) @overbalance